requirements.txt
pytest == 8.0.1
selenium == 4.18.1
allure-pytest == 2.13.2
faker == 23.2.1
pytest-xdist == 3.5.0
